Improved HDF5 I/O performance
With a different layout of the arrays in the datafiles the writing of
HDF5 files can now be done with ~ 7 GiB/s on the (busy) RZG Hydra.

This is achieved by writing the arrays in such a way that the sections
of each MPI-Task are stored in one contigous block in the resulting
file.

Still missing are the routines to read the files back in again.