Fix erroneous reading of HDF5 files
There was an error in the HDF5 read routine. I erroneously assumed that
the global arrays w, z, p, ... had the same LM numbering scheme as the
corresponding local arrays w_LMloc, z_LMloc, p_LMloc, etc., i.e. that
just w_LMloc == w(llm:ulm, :).

Therefore I assumed I could just call the HDF5 read routine on rank 0
with the global arrays instead of the local arrays and they would be
filled correctly.

Fixed by always reading the local arrays (for which the HDF5 routines
are meant to be used) and doing appropriate

   call gather_from_lo_to_rank0(w_LMloc(llm:,nR),w(1,nR))
   ...

calls that gather the local information back to task 0.

Of course, in the long run these global arrays should be eliminated
anyway, which is why I chose to do it this way instead of creating a
specific HDF5 read routine for the global arrays on rank 0, which would
avoid these MPI calls.