allow for an automatic detection of endianness
Now the endianness of the G_#.TAG and [V|B|T]_lmr_#.TAG files is automatically
detected. This implies that the G files produced by old version of openMPI (which
had no good support for "external32" representation is now working).

This also means that the environment variables F_UFMTENDIAN and
GFORTRAN_CONVERT_UNIT don't need to be set anymore in the sourceme.[c]sh files

	../../samples/boussBenchSat/B_coeff_rtest