port out_TO to local Simpson integration instead of the former spectral based integration.
Now portable to any setup: Finite Differences, Cheb with mapping.
Possible to tune the z and s resolution by using zDens and sDens parameters.
The viscous stress is found to be the most sensitive case to the resolution and sometimes
increasing zDens is mandatory to get a reasonable force balance.