new full sphere reference file
The change comes from the fact that now a vector transform readily gives the advection
terms. Beforehand, get_td recurrence relations were used to assemble those terms and no
care was taken to handle variable \ell there. As a consequence a small difference occured
in the benchmark which hence needed to be changed.