theta scrambling for Gauss's weights
This allows to suppress nThetaNHS in all diagnostics computed on the grid. This
also makes the code less dependent on the theta-scrambling. In principle, except
in a few places (like movie and outMisc where some tests on N/S hemispheres are
done), the code should be quite independent on the scrambling provided the definition
of theta, sin(theta) and so on are changed in horizontal.f90