bugfix in get_hit_times, l_correct_time
So far when specifying the sampling time for a dedicated output, like dt_movie,
the time array for the outputs was constructed assuming that the code would
run dt_old*n_time_steps, which could be quite wrong when the time step increases.

To prevent this, I rather fix the time array to one single entry in that case
and increment its value by dt_* whenever l_correct_time is hit.